+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - PR #3562 Fix 4GB limit for gzipped-compressed csv files
 - PR #3563 Use `__cuda_array_interface__` for serialization
 - PR #3548 Replaced CUDA_RT_CALL with CUDA_TRY
+- PR #3486 Pandas > 0.25 compatability 
 
 
 # cuDF 0.11.0 (11 Dec 2019)